+++ testbenchTbaGapp/src/test/java/de/gaalop/testbenchTbaGapp/gapp/Point3D.java
@@ -1,42 +1,42 @@
 package de.gaalop.testbenchTbaGapp.gapp;
 
 /**
- * Represents a point in a 3d space in float precision
+ * Represents a point in a 3d space in double precision
  * @author Christian Steinmetz
  */
 public class Point3D {
 
-    public float x;
-    public float y;
-    public float z;
+    public double x;
+    public double y;
+    public double z;
 
-    public Point3D(float x, float y, float z) {
+    public Point3D(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void setZ(float z) {
+    public void setZ(double z) {
         this.z = z;
     }
 }
